Fixed business add product description and added better handling to float. oroginally was a double.

+++ app/src/main/java/com/example/shopit/data/cart/CartProductDataClass.kt
@@ -3,7 +3,7 @@ package com.example.shopit.data.cart
 data class CartProductDataClass (
     var cartProductImage: String,   //URL
     var cartProductName: String,    //Name
-    var cartProductPrice: Double,   //Price
+    var cartProductPrice: Float,   //Price
     var cartProductQuantity: Int,
     var cartProductBarcode: String  //Barcode numbers
     )
